+++ persistence/test/fisherman_test.go
@@ -14,20 +14,20 @@ import (
 
 // TODO(andrew): Rename `addrBz` to `fishAddrBz` so tests are easier to read and understand. Ditto in all other locations.
 
-func FuzzFisherman(f *testing.F) {
+func fuzzFisherman(f *testing.F) {
 	fuzzSingleProtocolActor(f,
-		NewTestGenericActor(types.FishermanActor, newTestFisherman),
-		GetGenericActor(types.FishermanActor, getTestFisherman),
+		newTestGenericActor(types.FishermanActor, newTestFisherman),
+		getGenericActor(types.FishermanActor, getTestFisherman),
 		types.FishermanActor)
 }
 
 func TestGetSetFishermanStakeAmount(t *testing.T) {
-	db := NewTestPostgresContext(t, 1)
+	db := newTestPostgresContext(t, 1)
 	getTestGetSetStakeAmountTest(t, db, createAndInsertDefaultTestFisherman, db.GetFishermanStakeAmount, db.SetFishermanStakeAmount, 1)
 }
 
 func TestInsertFishermanAndExists(t *testing.T) {
-	db := NewTestPostgresContext(t, 0)
+	db := newTestPostgresContext(t, 0)
 
 	fisherman, err := createAndInsertDefaultTestFisherman(db)
 	require.NoError(t, err)
@@ -58,7 +58,7 @@ func TestInsertFishermanAndExists(t *testing.T) {
 }
 
 func TestUpdateFisherman(t *testing.T) {
-	db := NewTestPostgresContext(t, 0)
+	db := newTestPostgresContext(t, 0)
 
 	fisherman, err := createAndInsertDefaultTestFisherman(db)
 	require.NoError(t, err)
@@ -90,7 +90,7 @@ func TestUpdateFisherman(t *testing.T) {
 }
 
 func TestGetFishermenReadyToUnstake(t *testing.T) {
-	db := NewTestPostgresContext(t, 0)
+	db := newTestPostgresContext(t, 0)
 
 	fisherman, err := createAndInsertDefaultTestFisherman(db)
 	require.NoError(t, err)
@@ -132,7 +132,7 @@ func TestGetFishermenReadyToUnstake(t *testing.T) {
 }
 
 func TestGetFishermanStatus(t *testing.T) {
-	db := NewTestPostgresContext(t, 1)
+	db := newTestPostgresContext(t, 1)
 
 	fisherman, err := createAndInsertDefaultTestFisherman(db)
 	require.NoError(t, err)
@@ -152,7 +152,7 @@ func TestGetFishermanStatus(t *testing.T) {
 }
 
 func TestGetFishermanPauseHeightIfExists(t *testing.T) {
-	db := NewTestPostgresContext(t, 1)
+	db := newTestPostgresContext(t, 1)
 
 	fisherman, err := createAndInsertDefaultTestFisherman(db)
 	require.NoError(t, err)
@@ -172,7 +172,7 @@ func TestGetFishermanPauseHeightIfExists(t *testing.T) {
 }
 
 func TestSetFishermanPauseHeightAndUnstakeLater(t *testing.T) {
-	db := NewTestPostgresContext(t, 0)
+	db := newTestPostgresContext(t, 0)
 
 	fisherman, err := createAndInsertDefaultTestFisherman(db)
 	require.NoError(t, err)
@@ -199,7 +199,7 @@ func TestSetFishermanPauseHeightAndUnstakeLater(t *testing.T) {
 }
 
 func TestGetFishermanOutputAddress(t *testing.T) {
-	db := NewTestPostgresContext(t, 0)
+	db := newTestPostgresContext(t, 0)
 
 	fisherman, err := createAndInsertDefaultTestFisherman(db)
 	require.NoError(t, err)